third_party/blink/web_tests/virtual/partitioned-popins-disabled/external/wpt/partitioned-popins/partitioned-popins.localStorage.tentative.sub.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] Verify Partitioned Popins only have access to third-party localStorage
  assert_equals: expected "Found:ThirdParty-" but got "Found:FirstParty-"
Harness: the test ran to completion.
